Created and utilized the RoomFactory class.
